TMT
SYSTEM_SALT replaced with USER_SYSTEM_SALT, now  every user carry it in the DATA_FOLDER.
config salt backup removed completely, created 1 new function, create_user_system_salt()
this function called on create_user() it creates unique salts for every new user, old
initialize_user_system_salt() function changed and renamed. It now called in
authenticate_user() function to find correct .env file, it reads the salt from there
and sets it as USER_SYSTEM_SALT.
and more.

+++ SiriusVaultGUI/src/SiriusVaultFunctions.py
@@ -16,7 +16,6 @@ from dotenv import load_dotenv
 from threading import Timer
 from cryptography.fernet import Fernet
 
-load_dotenv()
 
 # Constants for user, password and vault management
 # GLOBAL PATH VARIABLES
@@ -99,87 +98,36 @@ def exit_program():
     sys.exit()
 
 # .env
-def set_file_readonly(filepath):
-    try: os.chmod(filepath, stat.S_IREAD)
+def set_file_readonly(USER_ENV):
+    try: os.chmod(USER_ENV, stat.S_IREAD)
     except: pass
 
-def remove_readonly(filepath):
-    try: os.chmod(filepath, stat.S_IWRITE)
+def remove_readonly(USER_ENV):
+    try: os.chmod(USER_ENV, stat.S_IWRITE)
     except: pass
 
-def initialize_system_salt():
-    env_path = os.path.join(BASE_DIR, ".env")
+def initialize_user_system_salt():
+    load_dotenv(USER_ENV)
     raw_env = os.getenv('SYSTEM_SALT')
     salt_from_env = raw_env if raw_env and raw_env.strip() else None
-    salt_from_config = None
-    if os.path.exists(CONFIG_FILE):
-        try:
-            with open(CONFIG_FILE, 'r') as f:
-                config_data = json.load(f)
-                salt_from_config = config_data.get("system_salt_backup")
-        except:
-            pass
-    
     final_salt = None
-
     if salt_from_env:
         final_salt = salt_from_env
-        if salt_from_config != salt_from_env:
-            if salt_from_config:
-                print(f"[WARNING] Salts not match. Using SALT from .env ({salt_from_env[:4]}...).")
-            try:
-                config_data = {}
-                if os.path.exists(CONFIG_FILE):
-                    with open(CONFIG_FILE, 'r') as f:
-                        try: config_data = json.load(f)
-                        except: config_data = {}
-                
-                config_data["system_salt_backup"] = final_salt
-                with open(CONFIG_FILE, 'w') as f:
-                    json.dump(config_data, f, indent=4)
-            except Exception as e:
-                print(f"[ERROR] Config backup failed: {e}")
-        set_file_readonly(env_path)
-    elif not salt_from_env and salt_from_config:
-        print("[WARNING] Missing .env file! Loading backup from config file...")
-        final_salt = salt_from_config
-        
-        try:
-            if os.path.exists(env_path): remove_readonly(env_path)
-            
-            with open(env_path, "w") as f:
-                f.write(f"SYSTEM_SALT={final_salt}")
-            
-            os.environ['SYSTEM_SALT'] = final_salt
-            set_file_readonly(env_path)
-        except Exception as e:
-            print(f"[ERROR] .env file cannot recovered: {e}")
-            sys.exit(1)
+        return final_salt
     else:
-        print("[WARNING] System cannot find the SALT. Creating a new one...")
-        new_salt = secrets.token_hex(16)
-        final_salt = new_salt
-        
-        if os.path.exists(env_path): remove_readonly(env_path)
-        with open(env_path, "w") as f:
+        print("[WARNING] System cannot find the SALT.")
+
+def create_user_system_salt():
+    
+    new_salt = secrets.token_hex(16)
+    if os.path.exists(USER_ENV): remove_readonly(USER_ENV)
+    try:
+        with open(USER_ENV, "w") as f:
             f.write(f"SYSTEM_SALT={new_salt}")
-        set_file_readonly(env_path)
-        
-        try:
-            config_data = {}
-            if os.path.exists(CONFIG_FILE):
-                with open(CONFIG_FILE, 'r') as f:
-                    try: config_data = json.load(f)
-                    except: config_data = {}
-            config_data["system_salt_backup"] = new_salt
-            with open(CONFIG_FILE, 'w') as f:
-                json.dump(config_data, f, indent=4)
-        except:
-            pass
-
-    return final_salt
-
-SYSTEM_SALT = initialize_system_salt()
+        set_file_readonly(USER_ENV)
+        print("User Salt created!")
+    except Exception as e:
+        print(f"User salt creation error: {e}")
 
 # CONFIG FUNC (TEST)
 def load_config():
@@ -237,7 +185,7 @@ def initialize_storage(target_path=None, default=True):
 
 def load_user_context(username):
     
-    global USER_DIR, VAULTS_DIR, RECOVERY_DIR, USER_DATA_FILE, ENC_USER_DATA_FILE, VAULT_METADATA_FILE, ENC_VAULT_METADATA_FILE, PASS_METADATA_FILE, ENC_PASS_METADATA_FILE
+    global USER_DIR, VAULTS_DIR, RECOVERY_DIR, USER_ENV, USER_DATA_FILE, ENC_USER_DATA_FILE, VAULT_METADATA_FILE, ENC_VAULT_METADATA_FILE, PASS_METADATA_FILE, ENC_PASS_METADATA_FILE
 
     user_hash = hashlib.sha256(username.encode('utf-8')).hexdigest()
     backup_name = username + "backup"
@@ -247,6 +195,7 @@ def load_user_context(username):
     VAULTS_DIR = os.path.join(USER_DIR, "Vaults")
 
     RECOVERY_DIR = os.path.join(USER_DIR, f"{backup_hash}")
+    USER_ENV = os.path.join(USER_DIR, ".env")
     USER_DATA_FILE = os.path.join(USER_DIR, "user.json")
     ENC_USER_DATA_FILE = os.path.join(USER_DIR, "user.json.enc")
     VAULT_METADATA_FILE = os.path.join(USER_DIR, "vault_metadata.json")
@@ -289,7 +238,7 @@ def decrypt_file(vault_key, encrypted_filepath, filename, destination_path):
 def encrypt_userdata_file(password):
     if not os.path.exists(USER_DATA_FILE):
         return
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -309,7 +258,7 @@ def encrypt_userdata_file(password):
 def encrypt_vaultdata_file(password):
     if not os.path.exists(VAULT_METADATA_FILE):
         return
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -329,7 +278,7 @@ def encrypt_vaultdata_file(password):
 def encrypt_passdata_file(password):
     if not os.path.exists(PASS_METADATA_FILE):
         return
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -350,7 +299,7 @@ def decrypt_userdata_file(password):
     if not os.path.exists(ENC_USER_DATA_FILE):
         return
     decrypted_path = USER_DATA_FILE
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -366,7 +315,7 @@ def decrypt_vaultdata_file(password):
     if not os.path.exists(ENC_VAULT_METADATA_FILE):
         return
     decrypted_path = VAULT_METADATA_FILE
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -382,7 +331,7 @@ def decrypt_passdata_file(password):
     if not os.path.exists(ENC_PASS_METADATA_FILE):
         return
     decrypted_path = PASS_METADATA_FILE
-    data_salt = bytes.fromhex(SYSTEM_SALT)
+    data_salt = bytes.fromhex(USER_SYSTEM_SALT)
     data_key = password
     jsonKey, _ = generate_key(data_key, data_salt)
     fernet = Fernet(jsonKey)
@@ -545,6 +494,7 @@ def format_drive_windows(drive_path):
 # User Management
 def create_user(username, password):
 
+    global USER_SYSTEM_SALT
     load_user_context(username)
 
     if os.path.exists(USER_DIR):
@@ -554,7 +504,9 @@ def create_user(username, password):
         os.makedirs(USER_DIR)
     if not os.path.exists(VAULTS_DIR):
         os.makedirs(VAULTS_DIR)
-    
+    create_user_system_salt()
+    USER_SYSTEM_SALT = initialize_user_system_salt()
+
     key, salt =generate_key(password)
     user_data = {"username": username,
                 "password_hash": key.hex(),
@@ -569,7 +521,9 @@ def create_user(username, password):
 
 def authenticate_user(username, password):
 
+    global USER_SYSTEM_SALT
     load_user_context(username)
+    USER_SYSTEM_SALT = initialize_user_system_salt()
     if not os.path.exists(ENC_USER_DATA_FILE):
         print("User not registered.")
         return False
@@ -596,24 +550,23 @@ def authenticate_user(username, password):
         return False
 
 # Delete user
-def delete_user():
-    username = session["authenticated_user"]
-    if not username: return
+def delete_user(username, user_password):
     load_user_context(username)
-    if os.path.exists(USER_DIR):
-        try:
-            shutil.rmtree(USER_DIR)
-            if not os.path.exists(USER_DIR):
-                print("User and all associated vaults deleted successfully!")
-                print("Good Bye!")
-                session["authenticated_user"] = None
-            else:
-                print("Something went wrong. Please be sure to close all files and try again.")
-                return
-        except Exception as e:
-            print(f"[ERROR] Cannot delete user: {e}")
-    else:
-        print("User not found.")
+    if authenticate_user(username, user_password):
+        if os.path.exists(USER_DIR):
+            try:
+                shutil.rmtree(USER_DIR)
+                if not os.path.exists(USER_DIR):
+                    print("User and all associated vaults deleted successfully!")
+                    print("Good Bye!")
+                    session["authenticated_user"] = None
+                else:
+                    print("Something went wrong. Please be sure to close all files and try again.")
+                    return
+            except Exception as e:
+                print(f"[ERROR] Cannot delete user: {e}")
+        else:
+            print("User not found.")
 
 # Vault Management
 def create_vault(vault_name, password): 
@@ -948,15 +901,19 @@ def remove_password_service(service_name, pass_Mngr=None):
 
 # Delete Password Manager
 def delete_passMngr(username, user_password):
+    load_user_context(username)
+    if authenticate_user(username, user_password):
+        if os.path.exists(PASS_METADATA_FILE):
+            os.remove(PASS_METADATA_FILE)
+        if os.path.exists(ENC_PASS_METADATA_FILE):
+            os.remove(ENC_PASS_METADATA_FILE)
 
-    if os.path.exists(PASS_METADATA_FILE):
-        os.remove(PASS_METADATA_FILE)
-    if os.path.exists(ENC_PASS_METADATA_FILE):
-        os.remove(ENC_PASS_METADATA_FILE)
-
-    if not os.path.exists(PASS_METADATA_FILE) and not os.path.exists(ENC_PASS_METADATA_FILE):
-        print("Password Manager deleted successfuly!")
-        return True
+        if not os.path.exists(PASS_METADATA_FILE) and not os.path.exists(ENC_PASS_METADATA_FILE):
+            print("Password Manager deleted successfuly!")
+            return True
+        else:
+            print("Something went wrong. Please be sure to close all files and try again.")
+            return False
     else:
-        print("Something went wrong. Please be sure to close all files and try again.")
+        print("User authentication failed.")
         return False
